+++ docs/source/changelog.rst
@@ -1,6 +1,18 @@
 Changelog
 ===============
 
+v0.1.19
+________
+
+* Fixed Copernicus search and download service
+
+v0.1.18
+________
+
+* Added Copernicus download service from
+  https://catalogue.dataspace.copernicus.eu
+  if copernicus_user and copernicus_password are provided.
+
 v0.1.17
 ________